Create lodash.js
This is a library that we use to implement some functions like cloneDeep,   that makes a deep copy of the board for example. a deep copy is when copied objects will have new references in memory and will not be related to the original objects. Therefore, any changes that we make to an object will not affect the other object.